+++ controlplane/doublezero-admin/src/cli/device.rs
@@ -1,23 +1,17 @@
 use clap::{Args, Subcommand};
-use doublezero_cli::{
-    allowlist::device::{
-        add::AddDeviceAllowlistCliCommand, list::ListDeviceAllowlistCliCommand,
-        remove::RemoveDeviceAllowlistCliCommand,
+use doublezero_cli::device::{
+    create::CreateDeviceCliCommand,
+    delete::DeleteDeviceCliCommand,
+    get::GetDeviceCliCommand,
+    interface::{
+        create::CreateDeviceInterfaceCliCommand, delete::DeleteDeviceInterfaceCliCommand,
+        get::GetDeviceInterfaceCliCommand, list::ListDeviceInterfaceCliCommand,
+        update::UpdateDeviceInterfaceCliCommand,
     },
-    device::{
-        create::CreateDeviceCliCommand,
-        delete::DeleteDeviceCliCommand,
-        get::GetDeviceCliCommand,
-        interface::{
-            create::CreateDeviceInterfaceCliCommand, delete::DeleteDeviceInterfaceCliCommand,
-            get::GetDeviceInterfaceCliCommand, list::ListDeviceInterfaceCliCommand,
-            update::UpdateDeviceInterfaceCliCommand,
-        },
-        list::ListDeviceCliCommand,
-        resume::ResumeDeviceCliCommand,
-        suspend::SuspendDeviceCliCommand,
-        update::UpdateDeviceCliCommand,
-    },
+    list::ListDeviceCliCommand,
+    resume::ResumeDeviceCliCommand,
+    suspend::SuspendDeviceCliCommand,
+    update::UpdateDeviceCliCommand,
 };
 
 #[derive(Debug, Subcommand)]
@@ -74,29 +68,7 @@ pub enum DeviceCommands {
     /// Delete a device
     #[clap()]
     Delete(DeleteDeviceCliCommand),
-    /// Manage device allowlist
-    #[clap()]
-    Allowlist(DeviceAllowlistCliCommand),
     /// Interface commands
     #[clap()]
     Interface(InterfaceCliCommand),
-}
-
-#[derive(Args, Debug)]
-pub struct DeviceAllowlistCliCommand {
-    #[command(subcommand)]
-    pub command: DeviceAllowlistCommands,
-}
-
-#[derive(Debug, Subcommand)]
-pub enum DeviceAllowlistCommands {
-    /// List device allowlist
-    #[clap()]
-    List(ListDeviceAllowlistCliCommand),
-    /// Add a device to the allowlist
-    #[clap()]
-    Add(AddDeviceAllowlistCliCommand),
-    /// Remove a device from the allowlist
-    #[clap()]
-    Remove(RemoveDeviceAllowlistCliCommand),
 }