+++ smartcontract/cli/src/doublezerocommand.rs
@@ -5,15 +5,9 @@ use doublezero_sdk::{
             close::CloseAccessPassCommand, get::GetAccessPassCommand, list::ListAccessPassCommand,
             set::SetAccessPassCommand,
         },
-        allowlist::{
-            device::{
-                add::AddDeviceAllowlistCommand, list::ListDeviceAllowlistCommand,
-                remove::RemoveDeviceAllowlistCommand,
-            },
-            foundation::{
-                add::AddFoundationAllowlistCommand, list::ListFoundationAllowlistCommand,
-                remove::RemoveFoundationAllowlistCommand,
-            },
+        allowlist::foundation::{
+            add::AddFoundationAllowlistCommand, list::ListFoundationAllowlistCommand,
+            remove::RemoveFoundationAllowlistCommand,
         },
         contributor::{
             create::CreateContributorCommand, delete::DeleteContributorCommand,
@@ -215,7 +209,6 @@ pub trait CliCommand {
         &self,
         cmd: ListFoundationAllowlistCommand,
     ) -> eyre::Result<Vec<Pubkey>>;
-    fn list_device_allowlist(&self, cmd: ListDeviceAllowlistCommand) -> eyre::Result<Vec<Pubkey>>;
     fn add_foundation_allowlist(
         &self,
         cmd: AddFoundationAllowlistCommand,
@@ -224,9 +217,6 @@ pub trait CliCommand {
         &self,
         cmd: RemoveFoundationAllowlistCommand,
     ) -> eyre::Result<Signature>;
-    fn add_device_allowlist(&self, cmd: AddDeviceAllowlistCommand) -> eyre::Result<Signature>;
-    fn remove_device_allowlist(&self, cmd: RemoveDeviceAllowlistCommand)
-        -> eyre::Result<Signature>;
     fn create_multicastgroup(
         &self,
         cmd: CreateMulticastGroupCommand,
@@ -547,9 +537,6 @@ impl CliCommand for CliCommandImpl<'_> {
     ) -> eyre::Result<Vec<Pubkey>> {
         cmd.execute(self.client)
     }
-    fn list_device_allowlist(&self, cmd: ListDeviceAllowlistCommand) -> eyre::Result<Vec<Pubkey>> {
-        cmd.execute(self.client)
-    }
     fn add_foundation_allowlist(
         &self,
         cmd: AddFoundationAllowlistCommand,
@@ -559,15 +546,6 @@ impl CliCommand for CliCommandImpl<'_> {
     fn remove_foundation_allowlist(
         &self,
         cmd: RemoveFoundationAllowlistCommand,
-    ) -> eyre::Result<Signature> {
-        cmd.execute(self.client)
-    }
-    fn add_device_allowlist(&self, cmd: AddDeviceAllowlistCommand) -> eyre::Result<Signature> {
-        cmd.execute(self.client)
-    }
-    fn remove_device_allowlist(
-        &self,
-        cmd: RemoveDeviceAllowlistCommand,
     ) -> eyre::Result<Signature> {
         cmd.execute(self.client)
     }